+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-sysctl CHANGELOG
+sysctl2 CHANGELOG
 =====================
 
-This file is used to list changes made in each version of the sysctl cookbook.
+This file is used to list changes made in each version of the sysctl2 cookbook.
 
 0.1.0
 -----
-- [michael.m.morris@gmail.com] - Initial release of sysctl cookbook
+- [michael.m.morris@gmail.com] - Initial release of sysctl2 cookbook